+++ NEWS.md
@@ -1,3 +1,10 @@
+## R package processR version 0.1.9
+===================================
+(2019-Apr-23)
+
+- New function condPlot2 added
+
+
 ## R package processR version 0.1.8.2
 =====================================
 (2019-Apr-16)